+++ artemis-gwt-test/src/test/java/com/artemis/AspectFieldHandlerTest.java
@@ -25,10 +25,12 @@ public class AspectFieldHandlerTest extends GWTTestCase {
 		AspectDescriptorPojo withAspectFields = new AspectDescriptorPojo();
 		world.inject(withAspectFields);
 
-//		assertEquals(reference, withAspectFields.ab);
+		assertEquals(reference, withAspectFields.ab);
 		assertNotNull(withAspectFields.aspect);
 		assertEquals(reference, withAspectFields.sub.getAspectBuilder());
 		assertNotNull(withAspectFields.transmuter);
+
+		checkArchetype(world, withAspectFields.archetype);
 	}
 
 
@@ -44,6 +46,14 @@ public class AspectFieldHandlerTest extends GWTTestCase {
 		assertNotNull(withAspectFields.aspect);
 		assertEquals(reference, withAspectFields.sub.getAspectBuilder());
 		assertNotNull(withAspectFields.transmuter);
+
+		checkArchetype(world, withAspectFields.archetype);
 	}
 
+	private static void checkArchetype(World world, Archetype archetype) {
+		Entity e = world.getEntity(world.create(archetype));
+		assertNotNull(e.getComponent(ComponentX.class));
+		assertNotNull(e.getComponent(ReusedComponent.class));
+		assertNull(e.getComponent(ComponentY.class));
+	}
 }