+++ artemis-gwt-test/src/test/java/com/artemis/system/AspectDescriptorSystem.java
@@ -1,36 +1,37 @@
 package com.artemis.system;
 
-import com.artemis.Aspect;
-import com.artemis.BaseSystem;
-import com.artemis.EntitySubscription;
-import com.artemis.EntityTransmuter;
+import com.artemis.*;
 import com.artemis.annotations.AspectDescriptor;
 import com.artemis.component.*;
 
 public class AspectDescriptorSystem extends BaseSystem {
-		@AspectDescriptor(
-			all = {ComponentX.class, ComponentY.class},
-			exclude = PooledString.class,
-			one = {ReusedComponent.class, EntityHolder.class})
-		public EntitySubscription sub;
+	@AspectDescriptor(
+		all = {ComponentX.class, ComponentY.class},
+		exclude = PooledString.class,
+		one = {ReusedComponent.class, EntityHolder.class})
+	public EntitySubscription sub;
 
-		@AspectDescriptor(
-			all = {ComponentX.class, ComponentY.class},
-			exclude = PooledString.class,
-			one = {ReusedComponent.class, EntityHolder.class})
-		public EntityTransmuter transmuter;
+	@AspectDescriptor(
+		all = {ComponentX.class, ComponentY.class},
+		exclude = PooledString.class,
+		one = {ReusedComponent.class, EntityHolder.class})
+	public EntityTransmuter transmuter;
 
-		@AspectDescriptor(
-			all = {ComponentX.class, ComponentY.class},
-			exclude = PooledString.class,
-			one = {ReusedComponent.class, EntityHolder.class})
-		public Aspect aspect;
+	@AspectDescriptor(
+		all = {ComponentX.class, ComponentY.class},
+		exclude = PooledString.class,
+		one = {ReusedComponent.class, EntityHolder.class})
+	public Aspect aspect;
 
-		@AspectDescriptor(
-			all = {ComponentX.class, ComponentY.class},
-			exclude = PooledString.class,
-			one = {ReusedComponent.class, EntityHolder.class})
-		public Aspect.Builder ab;
+	@AspectDescriptor(
+		all = {ComponentX.class, ComponentY.class},
+		exclude = PooledString.class,
+		one = {ReusedComponent.class, EntityHolder.class})
+	public Aspect.Builder ab;
+
+	@AspectDescriptor(
+		all = {ComponentX.class, ReusedComponent.class})
+	public Archetype archetype;
 
 	@Override
 	protected void processSystem() {}