+++ artemis/src/main/java/com/artemis/annotations/AspectDescriptor.java
@@ -6,15 +6,15 @@ import java.lang.annotation.*;
 
 /**
  * <p>Auto-configures fields pertaining to aspects. The annotated field
- * must be one the following types: {@link Aspect}, {@link Aspect.Builder},
+ * must be one the following types: {@link Archetype}, {@link Aspect}, {@link Aspect.Builder},
  * {@link EntitySubscription}, {@link EntityTransmuter}.</p>
  *
  * <p>This annotation works similar to {@link Wire}; fields are configured
  * during {@link EntitySystem#initialize()}, or explicitly via {@link World#inject(Object)}.</p>
  *
- * <h4>Note on EntityTransmuters</h4>
- * <p><code>all</code> and <code>exlude</code> coresspond to create/remove on transmuters.
- * Any types specified in <code>one</code> are ignored.</p>
+ * <h4>Note on EntityTransmuters/Archetypes</h4>
+ * <p><code>all</code> and <code>exclude</code> correspond to create/remove. Archetypes
+ * only feature create. Any types specified in <code>one</code> are ignored.</p>
  *
  * @see Wire
  */