+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 #### Version: 2.0.0-SNAPSHOT
 - **BREAKING CHANGES**
-  - Removed deprecated methods, see: 
+  - Removed deprecated code, see: [5de377d](https://github.com/junkdog/artemis-odb/commit/5de377dac1123285f5b80921d9f4fcdf34b26c05) 
 
 
 #### Version: 2.0.0-RC3 - 2016-06-21